created a seperate service file for insert logic and resolved PR comments
